Move /lib to /shared/lib folder in WebApp
