+++ application/account-management/WebApp/package.json
@@ -11,7 +11,7 @@
     "check": "biome check",
     "lint": "biome lint --write",
     "format": "biome format --write",
-    "swagger": "openapi-typescript lib/api/AccountManagement.Api.json -o lib/api/api.generated.d.ts",
+    "swagger": "openapi-typescript shared/lib/api/AccountManagement.Api.json -o shared/lib/api/api.generated.d.ts",
     "publish": "npm run build && mkdir -p ../Api/publish/WebApp/dist/ && rsync -r ./dist/ ../Api/publish/WebApp/dist/"
   },
   "dependencies": {
