+++ application/account-management/WebApp/shared/lib/applicationInsights/ApplicationInsightsProvider.tsx
