+++ application/account-management/WebApp/shared/lib/applicationInsights/config.tsx
