+++ application/account-management/WebApp/shared/lib/auth/AuthenticationProvider.tsx
