+++ application/account-management/WebApp/shared/translations/locale/da-DK.po
@@ -14,18 +14,6 @@ msgstr ""
 "Plural-Forms: \n"
 
 #. js-lingui-explicit-id
-msgid "Missing Fields. Failed to login."
-msgstr "Felter mangler. Kunne ikke logge ind."
-
-#. js-lingui-explicit-id
-msgid "Server error: Failed login."
-msgstr "Serverfejl: Kunne ikke logge ind."
-
-#. js-lingui-explicit-id
-msgid "Server error: Failed logout."
-msgstr "Serverfejl: Kunne ikke logge ud."
-
-#. js-lingui-explicit-id
 msgid "yourname@example.com"
 msgstr "ditnavn@example.com"
 
@@ -56,6 +44,18 @@ msgstr "Serverfejl: Kunne ikke færdigøre konto oprettelse."
 #. js-lingui-explicit-id
 msgid "subdomain"
 msgstr "subdomæne"
+
+#. js-lingui-explicit-id
+msgid "Missing Fields. Failed to login."
+msgstr "Felter mangler. Kunne ikke logge ind."
+
+#. js-lingui-explicit-id
+msgid "Server error: Failed login."
+msgstr "Serverfejl: Kunne ikke logge ind."
+
+#. js-lingui-explicit-id
+msgid "Server error: Failed logout."
+msgstr "Serverfejl: Kunne ikke logge ud."
 
 #. js-lingui-explicit-id
 msgid "Screenshots of the dashboard project showing mobile versions"
