+++ application/account-management/WebApp/shared/translations/locale/en-US.po
@@ -14,18 +14,6 @@ msgstr ""
 "Plural-Forms: \n"
 
 #. js-lingui-explicit-id
-msgid "Missing Fields. Failed to login."
-msgstr "Missing Fields. Failed to login."
-
-#. js-lingui-explicit-id
-msgid "Server error: Failed login."
-msgstr "Server error: Failed login."
-
-#. js-lingui-explicit-id
-msgid "Server error: Failed logout."
-msgstr "Server error: Failed logout."
-
-#. js-lingui-explicit-id
 msgid "yourname@example.com"
 msgstr "yourname@example.com"
 
@@ -56,6 +44,18 @@ msgstr "Server error: Failed to complete account registration."
 #. js-lingui-explicit-id
 msgid "subdomain"
 msgstr "subdomain"
+
+#. js-lingui-explicit-id
+msgid "Missing Fields. Failed to login."
+msgstr "Missing Fields. Failed to login."
+
+#. js-lingui-explicit-id
+msgid "Server error: Failed login."
+msgstr "Server error: Failed login."
+
+#. js-lingui-explicit-id
+msgid "Server error: Failed logout."
+msgstr "Server error: Failed logout."
 
 #. js-lingui-explicit-id
 msgid "Screenshots of the dashboard project showing mobile versions"
