+++ application/back-office/WebApp/shared/lib/applicationInsights/ApplicationInsightsProvider.tsx
