+++ application/back-office/WebApp/shared/lib/applicationInsights/config.tsx
