+++ application/back-office/WebApp/shared/lib/auth/AuthenticationProvider.tsx
